+++ src/Service/CodeCommit/src/Exception/OperationNotAllowedException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace AsyncAws\CodeCommit\Exception;
+
+use AsyncAws\Core\Exception\Http\ClientException;
+
+/**
+ * The requested action is not allowed.
+ */
+final class OperationNotAllowedException extends ClientException
+{
+}